refactor: break out into even more files
